Move getTxFee from account to test utils
